python-services/requirements.txt
# Core dependencies for environmental image analysis
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
scikit-image>=0.19.0
requests>=2.25.0
flask>=2.0.0
python-dotenv>=0.19.0

# Image processing and computer vision
scipy>=1.7.0

# Logging and utilities
loguru>=0.6.0

# Optional: TensorFlow for advanced ML (can be installed separately)
# tensorflow>=2.8.0